examples/applications/object_classification/kube_deployment/dataIngestionService/requirements.txt
paho-mqtt
numpy
pillow
requests
python-dotenv
boto3
qoa4ml